fxa_client
Enum FxaStateCheckerState

pub enum FxaStateCheckerState {
    GetAuthState,
    BeginOAuthFlow {
        scopes: Vec<String>,
        entrypoint: String,
    },
    BeginPairingFlow {
        pairing_url: String,
        scopes: Vec<String>,
        entrypoint: String,
    },
    CompleteOAuthFlow {
        code: String,
        state: String,
    },
    InitializeDevice,
    EnsureDeviceCapabilities,
    CheckAuthorizationStatus,
    GetProfile,
    Disconnect,
    Complete {
        new_state: FxaState,
    },
    Cancel,
}

State passed to the state checker, this is exactly the same as internal_machines::State except the Complete variant uses a named field for UniFFI compatibility.

Variants

GetAuthState

BeginOAuthFlow

Fields

scopes: Vec<String>

entrypoint: String

BeginPairingFlow

Fields

pairing_url: String

scopes: Vec<String>

entrypoint: String

CompleteOAuthFlow

Fields

code: String

state: String

InitializeDevice

EnsureDeviceCapabilities

CheckAuthorizationStatus

GetProfile

Disconnect

Complete

Fields

new_state: FxaState

Cancel