Function fxa_client::ffi::from_protobuf_ptr

pub unsafe fn from_protobuf_ptr<T, F: Message + Default + Into<T>>(
    data: *const u8, 
    len: i32
) -> Result<T>

Safety

data is a raw pointer to the protobuf data get_buffer will return an error if the length is invalid, or if the pointer is a null pointer