Module fxa_client::http_client

Structs

AuthorizationRequestParameters

Client

ClientConfigurationResponse

CommandData

DeviceLocation

DeviceResponseCommon

DeviceUpdateRequest
We use the double Option pattern in this struct. The outer option represents the existence of the field and the inner option its value or null. TL;DR: None: the field will not be present in the JSON body. Some(None): the field will have a null value. Some(Some(T)): the field will have the serialized value of T.

DeviceUpdateRequestBuilder

DuplicateTokenResponse

GetAttachedClientResponse

GetDeviceResponse

HawkRequestBuilder

IntrospectResponse

OAuthAuthResponse

OAuthTokenResponse

OpenIdConfigurationResponse

PendingCommand

PendingCommandsResponse

ProfileResponse

PushSubscription

ResponseAndETag

ScopedKeyDataResponse

Enums

DeviceType

HttpClientState

OAauthTokenRequest

Constants

HAWK_HKDF_SALT

HAWK_KEY_LENGTH

RETRY_AFTER_DEFAULT_SECONDS

Traits

FxAClient

Functions

bearer_token

derive_auth_key_from_session_token

fxa_client_configuration

get_keys_bundle

get_scoped_key_data_response

kw

openid_configuration

send_authorization_request

send_verification

Type Definitions

UpdateDeviceResponse