Trait fxa_client::internal::http_client::FxAClient

pub(crate) trait FxAClient {
    fn create_refresh_token_using_authorization_code(
        &self, 
        config: &Config, 
        code: &str, 
        code_verifier: &str
    ) -> Result<OAuthTokenResponse>;

    fn create_refresh_token_using_session_token(
        &self, 
        config: &Config, 
        session_token: &str, 
        scopes: &[&str]
    ) -> Result<OAuthTokenResponse>;

    fn check_refresh_token_status(
        &self, 
        config: &Config, 
        refresh_token: &str
    ) -> Result<IntrospectResponse>;

    fn create_access_token_using_refresh_token(
        &self, 
        config: &Config, 
        refresh_token: &str, 
        ttl: Option<u64>, 
        scopes: &[&str]
    ) -> Result<OAuthTokenResponse>;

    fn create_access_token_using_session_token(
        &self, 
        config: &Config, 
        session_token: &str, 
        scopes: &[&str]
    ) -> Result<OAuthTokenResponse>;

    fn create_authorization_code_using_session_token(
        &self, 
        config: &Config, 
        session_token: &str, 
        auth_params: AuthorizationRequestParameters
    ) -> Result<OAuthAuthResponse>;

    fn duplicate_session_token(
        &self, 
        config: &Config, 
        session_token: &str
    ) -> Result<DuplicateTokenResponse>;

    fn destroy_access_token(&self, config: &Config, token: &str) -> Result<()>;

    fn destroy_refresh_token(&self, config: &Config, token: &str) -> Result<()>;

    fn get_profile(
        &self, 
        config: &Config, 
        profile_access_token: &str, 
        etag: Option<String>
    ) -> Result<Option<ResponseAndETag<ProfileResponse>>>;

    fn get_pending_commands(
        &self, 
        config: &Config, 
        refresh_token: &str, 
        index: u64, 
        limit: Option<u64>
    ) -> Result<PendingCommandsResponse>;

    fn invoke_command(
        &self, 
        config: &Config, 
        refresh_token: &str, 
        command: &str, 
        target: &str, 
        payload: &Value
    ) -> Result<()>;

    fn update_device_record(
        &self, 
        config: &Config, 
        refresh_token: &str, 
        update: DeviceUpdateRequest<'_>
    ) -> Result<DeviceResponseCommon>;

    fn destroy_device_record(
        &self, 
        config: &Config, 
        refresh_token: &str, 
        id: &str
    ) -> Result<()>;

    fn get_devices(
        &self, 
        config: &Config, 
        refresh_token: &str
    ) -> Result<Vec<GetDeviceResponse>>;

    fn get_attached_clients(
        &self, 
        config: &Config, 
        session_token: &str
    ) -> Result<Vec<GetAttachedClientResponse>>;

    fn get_scoped_key_data(
        &self, 
        config: &Config, 
        session_token: &str, 
        client_id: &str, 
        scope: &str
    ) -> Result<HashMap<String, ScopedKeyDataResponse>>;

    fn get_fxa_client_configuration(
        &self, 
        config: &Config
    ) -> Result<ClientConfigurationResponse>;

    fn get_openid_configuration(
        &self, 
        config: &Config
    ) -> Result<OpenIdConfigurationResponse>;
}

Trait defining the low-level API for talking to the FxA server.

These are all the methods and datatypes used for interacting with the FxA server. It's defined as a trait mostly to make it mockable for testing purposes.

The default live implementation of this trait can be found in the Client struct, and you should consult that struct for documentation on specific methods.

A brief note on names: you'll see that the method names on this trait try to follow a pattern of <verb>_<noun>[_<additional info>](). Please try to keep to this pattern if you add methods to this trait.

Consistent names are helpful at the best of times, but they're particularly important here because so many of the nouns in OAuth contain embedded verbs! For example: at a glance, does a method named refresh_token refresh something called a "token", or does it return something called a "refresh token"? Using unambiguous verbs to start each method helps avoid confusion here.

Required methods

fn create_refresh_token_using_authorization_code(
&self,
config: &Config,
code: &str,
code_verifier: &str
) -> Result<OAuthTokenResponse>

fn create_refresh_token_using_session_token(
&self,
config: &Config,
session_token: &str,
scopes: &[&str]
) -> Result<OAuthTokenResponse>

fn check_refresh_token_status(
&self,
config: &Config,
refresh_token: &str
) -> Result<IntrospectResponse>

fn create_access_token_using_refresh_token(
&self,
config: &Config,
refresh_token: &str,
ttl: Option<u64>,
scopes: &[&str]
) -> Result<OAuthTokenResponse>

fn create_access_token_using_session_token(
&self,
config: &Config,
session_token: &str,
scopes: &[&str]
) -> Result<OAuthTokenResponse>

fn create_authorization_code_using_session_token(
&self,
config: &Config,
session_token: &str,
auth_params: AuthorizationRequestParameters
) -> Result<OAuthAuthResponse>

fn duplicate_session_token(
&self,
config: &Config,
session_token: &str
) -> Result<DuplicateTokenResponse>

fn destroy_access_token(&self, config: &Config, token: &str) -> Result<()>

fn destroy_refresh_token(&self, config: &Config, token: &str) -> Result<()>

fn get_profile(
&self,
config: &Config,
profile_access_token: &str,
etag: Option<String>
) -> Result<Option<ResponseAndETag<ProfileResponse>>>

fn get_pending_commands(
&self,
config: &Config,
refresh_token: &str,
index: u64,
limit: Option<u64>
) -> Result<PendingCommandsResponse>

fn invoke_command(
&self,
config: &Config,
refresh_token: &str,
command: &str,
target: &str,
payload: &Value
) -> Result<()>

fn update_device_record(
&self,
config: &Config,
refresh_token: &str,
update: DeviceUpdateRequest<'_>
) -> Result<DeviceResponseCommon>

fn destroy_device_record(
&self,
config: &Config,
refresh_token: &str,
id: &str
) -> Result<()>

fn get_devices(
&self,
config: &Config,
refresh_token: &str
) -> Result<Vec<GetDeviceResponse>>

fn get_attached_clients(
&self,
config: &Config,
session_token: &str
) -> Result<Vec<GetAttachedClientResponse>>

fn get_scoped_key_data(
&self,
config: &Config,
session_token: &str,
client_id: &str,
scope: &str
) -> Result<HashMap<String, ScopedKeyDataResponse>>

fn get_fxa_client_configuration(
&self,
config: &Config
) -> Result<ClientConfigurationResponse>

fn get_openid_configuration(
&self,
config: &Config
) -> Result<OpenIdConfigurationResponse>

Loading content...