Crate fxa_client

Re-exports

pub use crate::error::*;

Modules

auth

commands

config

device

error

ffi
This module implement the traits and some types that make the FFI code easier to manage.

http_client

migrator

msg_types

oauth

profile

push

scoped_keys

scopes

send_tab

state_persistence
This module implements the ability to serialize a FirefoxAccount struct to and from a JSON string. The idea is that calling code will use this to persist the account state to storage.

telemetry

util

Structs

AccessTokenInfo

CachedResponse

Config

FirefoxAccount

IntrospectInfo

Profile

RefreshToken

Enums

AccountEvent

IncomingDeviceCommand

Type Definitions

FxAClient